added an ability for community's sub-administrator (refs #228)

+++ lib/model/doctrine/CommunityMember.class.php
@@ -15,14 +15,34 @@ class CommunityMember extends BaseCommunityMember implements opAccessControlReco
     return $this->Community->generateRoleId($member);
   }
 
+  public function getPositions()
+  {
+    return Doctrine::getTable('CommunityMemberPosition')->getPositionsByMemberIdAndCommunityId($this->getMemberId(), $this->getCommunityId());
+  }
+
   public function hasPosition($name)
   {
-    return Doctrine::getTable('CommunityMemberPosition')->hasPosition($this->getMemberId(), $this->getCommunityId(), $name);
+    if (!is_array($name))
+    {
+      $name = array($name);
+    }
+    foreach ($name as $n)
+    {
+      if (in_array($n, $this->getPositions()))
+      {
+        return true;
+      }
+    }
+    return false;
   }
 
   public function addPosition($name)
   {
-    $object = Doctrine::getTable('CommunityMemberPosition')->findOneByCommunityMemberIdAndName($this->getId(), $name);
+    $object = null;
+    if (!$this->isNew())
+    {
+      $object = Doctrine::getTable('CommunityMemberPosition')->findOneByCommunityMemberIdAndName($this->getId(), $name);
+    }
     if (!$object)
     {
       $object = new CommunityMemberPosition();
@@ -36,11 +56,27 @@ class CommunityMember extends BaseCommunityMember implements opAccessControlReco
 
   public function removePosition($name)
   {
+    if ($this->isNew())
+    {
+      return false;
+    }
     $object = Doctrine::getTable('CommunityMemberPosition')->findOneByCommunityMemberIdAndName($this->getId(), $name);
     if (!$object)
     {
       throw new LogicException('The role data does not exist.');
     }
     $object->delete();
+  }
+
+  public function removeAllPosition()
+  {
+    if ($this->isNew())
+    {
+      return false;
+    }
+    Doctrine::getTable('CommunityMemberPosition')->createQuery()
+      ->where('community_member_id = ?', $this->getId())
+      ->delete()
+      ->execute();
   }
 }
